좌 클릭으로 오브젝트 둘 수 있는 기능 구현 1,2,3,4,5 키로 놓을 오브젝트 변경 가능 Tab 키로 프로세스 변경 가능(루프) 우 클릭으로 오브젝트 삭제할 수 있는 기능 구현 중

+++ Assets/New Script/ObjectsManager.cs
@@ -14,7 +14,7 @@ public class ObjectsManager: MonoBehaviour
 
 
     //유형에 따른 오브젝트 생성 함수
-    public GameObject InstantiateObject(Property.TaskType _taskType, int _taskID, int _x, int _y){
+    public GameObject InstantiateObject(Property.TaskType _taskType, int _taskID, float _x, float _y){
 
         GameObject _taskObject = null;
